Set number of guards to one.
Since August 2014, the parameter NumEntryGuards in consensus documents
is set to 1, effectively instructing Tor clients to use a single entry
guard.

Note that NumDirectoryGuards is still set to 3, but it doesn't matter
for TorPS as directory guards are not used to relay traffic, and are
therefore useless for correlation attacks.